+++ opusd.c
@@ -65,7 +65,7 @@ struct session {
   float deemph_gain;
   float deemph_state_left;
   float deemph_state_right;  
-  long long packets;
+  uint64_t packets;
 };
